+++ apps/dashboard/src/components/integrations/components/utils/helpers.ts
@@ -1,4 +1,11 @@
-import { ChatProviderIdEnum, EmailProviderIdEnum, SmsProviderIdEnum } from '@novu/shared';
+import {
+  ChatProviderIdEnum,
+  ConfigConfiguration,
+  CredentialsKeyEnum,
+  EmailProviderIdEnum,
+  IConfigCredential,
+  SmsProviderIdEnum,
+} from '@novu/shared';
 
 export function isDemoIntegration(providerId: string) {
   return (
@@ -6,4 +13,21 @@ export function isDemoIntegration(providerId: string) {
     providerId === SmsProviderIdEnum.Novu ||
     providerId === ChatProviderIdEnum.Novu
   );
+}
+
+export function configurationToCredential(config: ConfigConfiguration): IConfigCredential {
+  return {
+    key: config.key as CredentialsKeyEnum,
+    value: config.value,
+    placeholder: config.placeholder,
+    dropdown: config.dropdown,
+    displayName: config.displayName,
+    description: config.description,
+    type: config.type,
+    required: config.required,
+    links: config.links,
+    tooltip: {
+      text: config.tooltip,
+    },
+  } as IConfigCredential;
 }