+++ libs/internal-sdk/src/funcs/contextsRetrieve.ts
@@ -26,10 +26,11 @@ import { APICall, APIPromise } from '../types/async.js';
 import { Result } from '../types/fp.js';
 
 /**
- * Get context by id
+ * Retrieve a context
  *
  * @remarks
- * Retrieve a specific context by its type and id
+ * Retrieve a specific context by its type and id.
+ *       **type** and **id** are required fields, if the context does not exist, it returns the 404 response
  */
 export function contextsRetrieve(
   client: NovuCore,