+++ libs/internal-sdk/src/react-query/contextsDelete.ts
@@ -22,10 +22,11 @@ export type ContextsDeleteMutationVariables = {
 export type ContextsDeleteMutationData = operations.ContextsControllerDeleteContextResponse | undefined;
 
 /**
- * Delete context
+ * Delete a context
  *
  * @remarks
- * Delete a context by its type and id
+ * Delete a context by its type and id.
+ *       **type** and **id** are required fields, if the context does not exist, it returns the 404 response
  */
 export function useContextsDeleteMutation(
   options?: MutationHookOptions<ContextsDeleteMutationData, Error, ContextsDeleteMutationVariables>