+++ packages/shared/src/entities/integration/configuration.interface.ts
@@ -1,6 +1,13 @@
+export enum InboxCountTypeEnum {
+  NONE = 'none',
+  UNREAD = 'unread',
+  UNSEEN = 'unseen',
+}
+
 export interface IConfigurations {
   inboundWebhookEnabled?: boolean;
   inboundWebhookSigningKey?: string;
   configurationSetName?: string;
   pushResources?: string;
+  inboxCount?: InboxCountTypeEnum;
 }