+++ packages/shared/src/entities/integration/credential.interface.ts
@@ -47,5 +47,9 @@ export interface ICredentials {
   apiVersion?: string;
   appSid?: string;
   senderId?: string;
+  AppIOBaseUrl?: string;
+  AppIOSubscriptionId?: string;
+  AppIOBearerToken?: string;
+  AppIOOriginalSignature?: string;
   servicePlanId?: string;
 }